backend/requirements.txt
fastapi[all]
psycopg2-binary
python-dotenv
sqlalchemy